Added soybean to the GUI.
Soybean had an extra growth stage so all frames had to be modified to account for 16 growth stages. Other files had to handle this as well.
Added soybean to all the .dat files such as VarTypes, Emerge.dat...
Fixed an error in the HoftFarm weather file. Added Ames, Iowa weather file for soybeans.